#34 Implementação do método de cadastro de doação de item.

+++ src/main/java/br/ufpb/edoe/service/ItemService.java
@@ -1,14 +1,17 @@
 package br.ufpb.edoe.service;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import br.ufpb.edoe.dto.DonateDTO;
 import br.ufpb.edoe.dto.ItemDTO;
 import br.ufpb.edoe.dto.UpdateItemRequestDTO;
+import br.ufpb.edoe.entity.Donation;
 import br.ufpb.edoe.entity.Item;
 import br.ufpb.edoe.entity.User;
 import br.ufpb.edoe.enums.UserRoleType;
@@ -17,6 +20,7 @@ import br.ufpb.edoe.exceptions.descriptor.DescriptorNotFoundException;
 import br.ufpb.edoe.exceptions.item.ItemNotFoundException;
 import br.ufpb.edoe.exceptions.user.InvalidUserRoleException;
 import br.ufpb.edoe.exceptions.user.UserNotLoggedException;
+import br.ufpb.edoe.repository.DonationRepository;
 import br.ufpb.edoe.repository.ItemRepository;
 import br.ufpb.edoe.repository.UserRepository;
 import br.ufpb.edoe.security.JWTSecurity;
@@ -29,6 +33,9 @@ public class ItemService {
 
     @Autowired
     private ItemRepository repository;
+
+    @Autowired
+    private DonationRepository donateRepository;
 
     @Autowired
     private DescriptorService descriptorService;
@@ -47,8 +54,16 @@ public class ItemService {
         }
 
         Optional<User> user = userRepository.findByEmail(loggedEmail.get());
-        if (!user.get().getPapel().equals(UserRoleType.APENAS_RECEPTOR) && item.getIsRequired() == IS_REQUIRED) {
-            throw new InvalidUserRoleException("Usuário não possui o papel de APENAS_RECEPTOR", "UserService.addItem");
+        if (!user.get().getPapel().equals(UserRoleType.APENAS_RECEPTOR) && item.getIsRequired() == 1) {
+            throw new InvalidUserRoleException(
+                    "Usuário não possui o papel de APENAS_RECEPTOR para cadastrar um item necessário.",
+                    "ItemService.addItem");
+        }
+
+        if (user.get().getPapel().equals(UserRoleType.APENAS_RECEPTOR) && item.getIsRequired() == 0) {
+            throw new InvalidUserRoleException(
+                    "Usuário não possui o papel de APENAS_DOADOR, DOADOR_RECEPTOR ou ADMIN para cadastrar um item a ser doado.",
+                    "ItemService.addItem");
         }
 
         if (!checkItemsIsRequired(item.getIsRequired())) {
@@ -72,7 +87,7 @@ public class ItemService {
                     "ItemService.getItemsByDescriptorId");
         }
         repository.findAllByDescriptorId(id).forEach(item -> {
-            if (item.getIsRequired() == type)
+            if (item.getIsRequired() == type && item.getIsRemoved() == 0)
                 listItems.add(new ItemDTO(item));
         });
         return listItems;
@@ -85,7 +100,7 @@ public class ItemService {
                     "ItemService.getTop10ItemsByQty");
         }
         repository.findTop10ByOrderByQtyDesc().forEach(item -> {
-            if (item.getIsRequired() == type)
+            if (item.getIsRequired() == type && item.getIsRemoved() == 0)
                 listItems.add(new ItemDTO(item));
         });
         return listItems;
@@ -98,14 +113,14 @@ public class ItemService {
                     "ItemService.getItemsByString");
         }
         repository.findAllByDescriptorNameContainingIgnoreCase(search).forEach(item -> {
-            if (item.getIsRequired() == type)
+            if (item.getIsRequired() == type && item.getIsRemoved() == 0)
                 listItems.add(new ItemDTO(item));
         });
         listItems.sort((item1, item2) -> item1.getDescriptor().getName().compareTo(item2.getDescriptor().getName()));
         return listItems;
     }
 
-    public List<ItemDTO> getItemsMatches(int id, String header){
+    public List<ItemDTO> getItemsMatches(int id, String header) {
         Item item = this.repository.findById(id).get();
 
         Optional<String> loggedEmail = jwtSecurity.getUser(header);
@@ -115,14 +130,15 @@ public class ItemService {
 
         Optional<User> user = userRepository.findByEmail(loggedEmail.get());
         if (!user.get().getPapel().equals(UserRoleType.APENAS_RECEPTOR)) {
-            throw new InvalidUserRoleException("Usuário não possui o papel de APENAS_RECEPTOR", "UserService.getItemsMatches");
+            throw new InvalidUserRoleException("Usuário não possui o papel de APENAS_RECEPTOR",
+                    "UserService.getItemsMatches");
         }
 
         if (item.getIsRequired() != IS_REQUIRED) {
             throw new BadRequestParamsException("Parâmetro 'required' inválido.", "ItemService.getItemsMatches");
         }
 
-        return this.getItemsByDescriptorId(item.getDescriptor().getId(), DONATE);            
+        return this.getItemsByDescriptorId(item.getDescriptor().getId(), DONATE);
     }
 
     public ItemDTO removeItem(int id, String header) {
@@ -135,11 +151,22 @@ public class ItemService {
 
         Optional<User> user = userRepository.findByEmail(loggedEmail.get());
         if (!user.get().getPapel().equals(UserRoleType.APENAS_RECEPTOR) && i.getIsRequired() == 1) {
-            throw new InvalidUserRoleException("Usuário não possui o papel de APENAS_RECEPTOR",
+            throw new InvalidUserRoleException(
+                    "Usuário não possui o papel de APENAS_RECEPTOR para remover um item necessário.",
                     "UserService.removeItem");
         }
 
-        this.repository.delete(i);
+        if (user.get().getPapel().equals(UserRoleType.APENAS_RECEPTOR) && i.getIsRequired() == 0) {
+            throw new InvalidUserRoleException(
+                    "Usuário não possui o papel de APENAS_DOADOR, DOADOR_RECEPTOR ou ADMIN para remover um item a ser doado.",
+                    "ItemService.addItem");
+        }
+
+        if (i.getIsRemoved() == 0) {
+            i.setIsRemoved(1);
+            this.repository.delete(i);
+        }
+
         return new ItemDTO(i);
     }
 
@@ -170,7 +197,67 @@ public class ItemService {
         return new ItemDTO(i.get());
     }
 
+    public void donateItem(DonateDTO donateDTO, String header) {
+        Optional<Item> itemDonate = repository.findById(donateDTO.getItemDonateId());
+        Optional<Item> itemRequired = repository.findById(donateDTO.getItemRequiredId());
+
+        if (!itemDonate.isPresent() || itemDonate.get().getIsRemoved() == 1) {
+            throw new ItemNotFoundException("Id de Item para doação não encontrado!", "ItemService.donateItem");
+        }
+        if (!itemRequired.isPresent() || itemRequired.get().getIsRemoved() == 1) {
+            throw new ItemNotFoundException("Id de Item necessário não encontrado!", "ItemService.donateItem");
+        }
+
+        Optional<String> loggedEmail = jwtSecurity.getUser(header);
+        if (!loggedEmail.isPresent()) {
+            throw new UserNotLoggedException("Usuário não logado", "ItemService.donateItem");
+        }
+
+        Optional<User> user = userRepository.findByEmail(loggedEmail.get());
+        if (!user.get().getPapel().equals(UserRoleType.APENAS_DOADOR)) {
+            throw new InvalidUserRoleException("Usuário não possui o papel de APENAS_DOADOR", "ItemService.donateItem");
+        }
+
+        if (!validateDonation(donateDTO)) {
+            throw new BadRequestParamsException("Itens com descritores divergentes", "ItemService.updateItem");
+        }
+
+        User userReceptor = userRepository.findByEmail(itemRequired.get().getAssociatedEmail()).get();
+
+        int itemDonateQtyUpdate = itemDonate.get().getQty() - donateDTO.getQty();
+        int itemRequiredQtyUpdate = itemRequired.get().getQty() - donateDTO.getQty();
+
+        // Verifica se o item doado ou requirido deve ser removido
+        if (itemDonateQtyUpdate == 0)
+            itemDonate.get().setIsRemoved(1);
+        if (itemRequiredQtyUpdate == 0)
+            itemRequired.get().setIsRemoved(1);
+
+        // Atualizando a quant de iten doados e necessarios
+        itemDonate.get().setQty(itemDonateQtyUpdate);
+        itemRequired.get().setQty(itemRequiredQtyUpdate);
+
+        // Atualizando os itens em banco
+        List<Item> items = Arrays.asList(itemDonate.get(), itemRequired.get());
+        repository.saveAllAndFlush(items);
+
+        // Criando doacao
+        Donation donate = new Donation();
+        donate.setItemDonate(itemDonate.get());
+        donate.setUserDonate(user.get());
+        donate.setUserReceptor(userReceptor);
+
+        // Salvando doacao em banco
+        donateRepository.save(donate);
+    }
+
     public boolean checkItemsIsRequired(int required) {
         return required != DONATE && required != IS_REQUIRED ? false : true;
+    }
+
+    private boolean validateDonation(DonateDTO donateDTO) {
+        Item itemDonate = repository.getById(donateDTO.getItemDonateId());
+        Item itemRequired = repository.getById(donateDTO.getItemRequiredId());
+        return itemDonate.getDescriptor().getId() == itemRequired.getDescriptor().getId();
     }
 }
